+++ example/echo/echo_server.cc
@@ -2,10 +2,17 @@
 
 using namespace kanon;
 
-int main() {
+int main(int argc, char **argv) {
   EventLoop loop;
-
+  
   EchoServer server(loop);
+  if (argc > 1) {
+    if (strcmp(argv[1], "1") == 0) {
+      LOG_INFO << "Pool is enabled";
+      server.SetChunkPerBlock(10000);
+      server.EnablePool(true);
+    }
+  }
 
   server.StartRun();
   loop.StartLoop();